edited process for taking in values
Instead of using variables rowcounter and column counter I used i's and j's and it worked the same, but with less processing time and power required (if not I will leave a comment of the original for loop above one).

However because of the for loop formatting, whenever there is a a column or row equal to 1, whereas the other is not equal to one, the comes a problem (most likely due to for loop formatting).